feat: rewrite integration to be SpanProcessor
