+++ gradle.properties
@@ -1,2 +1,2 @@
 pyroscope_version=0.9.1
-otel_profiling_version=0.9.1
+otel_profiling_version=0.9.1-SNAPSHOT
